+++ lib/capybara/poltergeist/driver.rb
@@ -165,8 +165,8 @@ module Capybara::Poltergeist
       browser.set_paper_size(size)
     end
 
-    def zoom_factor=(zoom_factor=nil)
-      browser.set_zoom_factor(zoom_factor) if zoom_factor
+    def zoom_factor=(zoom_factor)
+      browser.set_zoom_factor(zoom_factor)
     end
 
     def resize(width, height)