+++ lib/capybara/poltergeist/node.rb
@@ -70,6 +70,7 @@ module Capybara::Poltergeist
       elsif tag_name == 'textarea'
         command :set, value.to_s
       elsif self[:contenteditable] == 'true'
+        click
         command :delete_text
         send_keys(value.to_s)
       end