+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 #### Bug fixes ####
 
 *   Fix `set` appending to `contenteditable` instead of replacing its text
-    (Erik Ostrom) [Issue #432]
+    (Pedro Carriço and Erik Ostrom) [Issue #432]
 
 ### 1.5.0 ###